+++ app/admin/members/components/DataCenterTab.tsx
@@ -2,163 +2,9 @@
 
 import { useState, useRef } from 'react'
 import { useLanguage } from '@/lib/hooks/useLanguage'
-
-// ── Types ─────────────────────────────────────────────────────────────────────
-
-type NewMember   = { abo_number: string; name: string; abo_level: string }
-type LevelChange = { abo_number: string; name: string; prev_level: string; new_level: string }
-type BonusChange = { abo_number: string; name: string; prev_bonus: number; new_bonus: number }
-
-type UnrecognizedRow = {
-  abo_number: string
-  name: string
-  sponsor_abo_number: string | null
-}
-
-type ManualMemberNoAbo = {
-  id: string
-  first_name: string
-  last_name: string
-  upline_abo_number: string | null
-}
-
-type ImportResult = {
-  inserted: number
-  errors: { abo_number: string; error: string }[]
-  diff: {
-    new_members:   NewMember[]
-    level_changes: LevelChange[]
-    bonus_changes: BonusChange[]
-  }
-  unrecognized: UnrecognizedRow[]
-  manual_members_no_abo: ManualMemberNoAbo[]
-}
-
-// ── CSV helpers ───────────────────────────────────────────────────────────────
-
-const HEADER_MAP: Record<string, string> = {
-  'ABO Level': 'abo_level', 'Sponsor ABO Number': 'sponsor_abo_number',
-  'ABO Number': 'abo_number', 'Country': 'country', 'Name': 'name',
-  'Entry Date': 'entry_date', 'Phone': 'phone', 'Email': 'email',
-  'Address': 'address', 'Renewal Date': 'renewal_date', 'GPV': 'gpv',
-  'PPV': 'ppv', 'Bonus Percent': 'bonus_percent', 'GBV': 'gbv',
-  'Customer PV': 'customer_pv', 'Ruby PV': 'ruby_pv', 'Customers': 'customers',
-  ' Customers': 'customers', 'Points to Next level': 'points_to_next_level',
-  'Qualified Legs': 'qualified_legs', 'Group Size': 'group_size',
-  'Personal Order Count': 'personal_order_count', 'Group Orders Count': 'group_orders_count',
-  'Sponsoring': 'sponsoring', 'Annual PPV': 'annual_ppv',
-}
-
-const HEADER_MAP_BG: Record<string, string> = {
-  'Ниво на СБА': 'abo_level',
-  'Номер на Спонсориращия СБА': 'sponsor_abo_number',
-  'Номер на СБА': 'abo_number',
-  'Държава': 'country',
-  'Име': 'name',
-  'Дата на въвеждане': 'entry_date',
-  'Телефон': 'phone',
-  'Електронна поща': 'email',
-  'Адрес': 'address',
-  'Дата на подновяване': 'renewal_date',
-  'ГТС': 'gpv',
-  'ЛТС': 'ppv',
-  'Процент на възнаграждение': 'bonus_percent',
-  'ГБО': 'gbv',
-  'Клиентска ТС': 'customer_pv',
-  'ТС за Рубин': 'ruby_pv',
-  'Клиенти': 'customers',
-  'Точки до следващото ниво': 'points_to_next_level',
-  'Квалифицирани звена': 'qualified_legs',
-  'Размер на група': 'group_size',
-  'Брой лични поръчки': 'personal_order_count',
-  'Брой групови поръчки': 'group_orders_count',
-  'Спонсориране': 'sponsoring',
-  'Годишна ЛТС:': 'annual_ppv',
-  'ТС общо за организацията': 'org_total_pv',
-}
-
-const NUMERIC_KEYS = new Set([
-  'gpv', 'ppv', 'bonus_percent', 'gbv', 'customer_pv', 'ruby_pv',
-  'customers', 'points_to_next_level', 'qualified_legs', 'group_size',
-  'personal_order_count', 'group_orders_count', 'sponsoring', 'annual_ppv',
-])
-
-function sanitizeNumeric(val: string, isBG: boolean): string {
-  if (isBG) {
-    return val.replace(/[\u00A0\s]/g, '').replace(/,/g, '.')
-  }
-  return val.replace(/,/g, '')
-}
-
-const MONTH_MAP: Record<string, string> = {
-  January: '01', February: '02', March: '03', April: '04',
-  May: '05', June: '06', July: '07', August: '08',
-  September: '09', October: '10', November: '11', December: '12',
-}
-
-function parseDate(val: string): string {
-  if (!val) return ''
-  if (/^\d{4}-\d{2}-\d{2}$/.test(val)) return val
-  const m = val.match(/^(\d{1,2})\s+([A-Za-z]+)\s+(\d{4})$/)
-  if (m) {
-    const month = MONTH_MAP[m[2]]
-    if (month) return `${m[3]}-${month}-${m[1].padStart(2, '0')}`
-  }
-  return ''
-}
-
-function splitCSVLine(line: string): string[] {
-  const result: string[] = []
-  let current = ''
-  let inQuotes = false
-  for (let i = 0; i < line.length; i++) {
-    const ch = line[i]
-    if (ch === '"') {
-      if (inQuotes && line[i + 1] === '"') { current += '"'; i++ }
-      else inQuotes = !inQuotes
-    } else if (ch === ',' && !inQuotes) {
-      result.push(current.trim())
-      current = ''
-    } else {
-      current += ch
-    }
-  }
-  result.push(current.trim())
-  return result
-}
-
-function parseCSV(text: string): Record<string, string>[] {
-  const cleaned = text.replace(/^\uFEFF/, '')
-  const allLines = cleaned.trim().split('\n')
-
-  const headerIdx = allLines.findIndex(
-    l => l.includes('ABO Number') || l.includes('Номер на СБА')
-  )
-  if (headerIdx === -1) return []
-
-  const isBG = allLines[headerIdx].includes('Номер на СБА')
-  const activeMap = isBG ? HEADER_MAP_BG : HEADER_MAP
-
-  const dataLines = allLines.slice(headerIdx)
-  const headers = splitCSVLine(dataLines[0])
-
-  return dataLines.slice(1)
-    .filter(line => line.trim() !== '')
-    .map(line => {
-      const values = splitCSVLine(line)
-      const row: Record<string, string> = {}
-      headers.forEach((h, i) => {
-        const dbKey = activeMap[h] ?? h.toLowerCase().replace(/\s+/g, '_')
-        let val = values[i] ?? ''
-        if (dbKey === 'bonus_percent') val = val.replace('%', '').trim()
-        if (dbKey === 'entry_date' || dbKey === 'renewal_date') val = parseDate(val)
-        if (dbKey === 'phone') val = val.replace(/^Primary:\s*/i, '').trim()
-        if (NUMERIC_KEYS.has(dbKey)) val = sanitizeNumeric(val, isBG)
-        row[dbKey] = val
-      })
-      return row
-    })
-}
+import { apiClient } from '@/lib/apiClient'
+import { parseCSV, type ImportResult } from '@/lib/csv-import'
+import { ReconciliationPanel } from './ReconciliationPanel'
 
 // ── DiffSection ───────────────────────────────────────────────────────────────
 
@@ -181,119 +27,6 @@ function DiffSection({
         </svg>
       </button>
       {open && <div className="mt-2 space-y-1">{children}</div>}
-    </div>
-  )
-}
-
-// ── ReconciliationPanel ───────────────────────────────────────────────────────
-
-function ReconciliationPanel({
-  initialUnrecognized, initialProfiles,
-}: { initialUnrecognized: UnrecognizedRow[]; initialProfiles: ManualMemberNoAbo[] }) {
-  const { t } = useLanguage()
-  const [unrecognized, setUnrecognized] = useState(initialUnrecognized)
-  const [profiles, setProfiles] = useState(initialProfiles)
-  const [selectedLos, setSelectedLos] = useState<string | null>(null)
-  const [selectedProfileId, setSelectedProfileId] = useState<string | null>(null)
-  const [linking, setLinking] = useState(false)
-  const [linkError, setLinkError] = useState<string | null>(null)
-
-  if (unrecognized.length === 0 && profiles.length === 0) return null
-
-  const selectedLosRow = unrecognized.find(r => r.abo_number === selectedLos) ?? null
-
-  async function handleLink() {
-    if (!selectedProfileId || !selectedLos || !selectedLosRow) return
-    setLinking(true)
-    setLinkError(null)
-    try {
-      const res = await fetch(`/api/admin/members/${selectedProfileId}/assign-abo`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ abo_number: selectedLosRow.abo_number, sponsor_abo_number: selectedLosRow.sponsor_abo_number ?? null }),
-      })
-      const data = await res.json()
-      if (!res.ok) throw new Error(data.error ?? 'Link failed')
-      setUnrecognized(prev => prev.filter(r => r.abo_number !== selectedLos))
-      setProfiles(prev => prev.filter(p => p.id !== selectedProfileId))
-      setSelectedLos(null)
-      setSelectedProfileId(null)
-    } catch (err: unknown) {
-      setLinkError(err instanceof Error ? err.message : 'Link failed')
-    } finally {
-      setLinking(false)
-    }
-  }
-
-  return (
-    <div className="mt-6 p-5 rounded-2xl border" style={{ borderColor: 'var(--border-default)', backgroundColor: 'var(--bg-card)' }}>
-      <div className="flex items-start justify-between gap-4 mb-4">
-        <div>
-          <p className="text-sm font-semibold" style={{ color: 'var(--text-primary)' }}>Reconciliation</p>
-          <p className="text-xs mt-0.5" style={{ color: 'var(--text-secondary)' }}>{t('admin.data.reconciliation.desc')}</p>
-        </div>
-        {selectedLos && selectedProfileId && (
-          <button onClick={handleLink} disabled={linking}
-            className="flex-shrink-0 px-4 py-2 rounded-lg text-sm font-semibold text-white disabled:opacity-50"
-            style={{ backgroundColor: 'var(--brand-teal)' }}>
-            {linking ? 'Linking...' : 'Link'}
-          </button>
-        )}
-      </div>
-      {linkError && <p className="text-xs mb-3" style={{ color: 'var(--brand-crimson)' }}>{linkError}</p>}
-      <div className="grid grid-cols-2 gap-4">
-        <div>
-          <p className="text-xs font-semibold tracking-widest uppercase mb-2" style={{ color: 'var(--text-secondary)' }}>
-            Unrecognized in LOS — {unrecognized.length}
-          </p>
-          {unrecognized.length === 0 ? (
-            <p className="text-xs" style={{ color: 'var(--text-secondary)' }}>{t('admin.data.reconciliation.allMatched')}</p>
-          ) : (
-            <div className="space-y-1.5">
-              {unrecognized.map(row => (
-                <button key={row.abo_number}
-                  onClick={() => setSelectedLos(s => s === row.abo_number ? null : row.abo_number)}
-                  className="w-full text-left px-3 py-2.5 rounded-xl text-xs transition-colors"
-                  style={{
-                    backgroundColor: selectedLos === row.abo_number ? 'rgba(62,119,133,0.15)' : 'rgba(0,0,0,0.03)',
-                    border: `1px solid ${selectedLos === row.abo_number ? '#3E7785' : 'transparent'}`,
-                  }}>
-                  <span className="font-mono font-medium block" style={{ color: 'var(--text-primary)' }}>{row.abo_number}</span>
-                  <span style={{ color: 'var(--text-secondary)' }}>{row.name}</span>
-                  {row.sponsor_abo_number && (
-                    <span className="block mt-0.5" style={{ color: 'var(--text-secondary)', opacity: 0.6 }}>Sponsor {row.sponsor_abo_number}</span>
-                  )}
-                </button>
-              ))}
-            </div>
-          )}
-        </div>
-        <div>
-          <p className="text-xs font-semibold tracking-widest uppercase mb-2" style={{ color: 'var(--text-secondary)' }}>
-            No ABO — awaiting position — {profiles.length}
-          </p>
-          {profiles.length === 0 ? (
-            <p className="text-xs" style={{ color: 'var(--text-secondary)' }}>{t('admin.data.reconciliation.noManualMembers')}</p>
-          ) : (
-            <div className="space-y-1.5">
-              {profiles.map(p => (
-                <button key={p.id}
-                  onClick={() => setSelectedProfileId(s => s === p.id ? null : p.id)}
-                  className="w-full text-left px-3 py-2.5 rounded-xl text-xs transition-colors"
-                  style={{
-                    backgroundColor: selectedProfileId === p.id ? 'rgba(62,119,133,0.15)' : 'rgba(0,0,0,0.03)',
-                    border: `1px solid ${selectedProfileId === p.id ? '#3E7785' : 'transparent'}`,
-                  }}>
-                  <span className="font-medium block" style={{ color: 'var(--text-primary)' }}>{p.first_name} {p.last_name}</span>
-                  {p.upline_abo_number && (
-                    <span style={{ color: 'var(--text-secondary)' }}>Upline {p.upline_abo_number}</span>
-                  )}
-                </button>
-              ))}
-            </div>
-          )}
-        </div>
-      </div>
     </div>
   )
 }
@@ -338,13 +71,10 @@ export function DataCenterTab() {
     setError(null)
     setResult(null)
     try {
-      const res = await fetch('/api/admin/los-import', {
+      const data = await apiClient<ImportResult>('/api/admin/los-import', {
         method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({ rows: parsedRows }),
       })
-      const data = await res.json()
-      if (!res.ok) throw new Error(data.error ?? 'Import failed')
       setResult(data)
     } catch (err: unknown) {
       setError(err instanceof Error ? err.message : 'Unknown error')
